Use int64 for blockheight

+++ bifrostv2/pkg/blockclients/bnb/bnb.go
@@ -26,7 +26,7 @@ import (
 type Client struct {
 	client                   *rpc.HTTP
 	cfg                      config.ChainConfigurations
-	lastScannedBlockHeight   uint64
+	lastScannedBlockHeight   int64
 	fnLastScannedBlockHeight types.FnLastScannedBlockHeight
 	logger                   zerolog.Logger
 	backOffCtrl              backoff.ExponentialBackOff
@@ -65,7 +65,7 @@ func setNetwork(cfg config.ChainConfigurations) btypes.ChainNetwork {
 	return network
 }
 
-func (c *Client) getBlock(blockHeight uint64) (*ctypes.ResultBlock, error) {
+func (c *Client) getBlock(blockHeight int64) (*ctypes.ResultBlock, error) {
 	return c.client.Block(pointy.Int64(int64(blockHeight)))
 }
 
@@ -96,7 +96,7 @@ func (c *Client) scanBlocks(blockInChan chan<- types.Block) {
 		block, err := c.getBlock(c.lastScannedBlockHeight)
 		if err != nil || block.Block == nil {
 			d := c.backOffCtrl.NextBackOff()
-			c.logger.Error().Err(err).Uint64("lastScannedBlockHeight", c.lastScannedBlockHeight).Str("backOffCtrl", d.String()).Msg("getBlock failed")
+			c.logger.Error().Err(err).Int64("lastScannedBlockHeight", c.lastScannedBlockHeight).Str("backOffCtrl", d.String()).Msg("getBlock failed")
 			time.Sleep(d)
 			continue
 		}
@@ -112,7 +112,7 @@ func (c *Client) processBlock(block *ctypes.ResultBlock) types.Block {
 
 	b.Chain = common.BNBChain
 	b.BlockHash = block.Block.Hash().String()
-	b.BlockHeight = uint64(block.Block.Header.Height)
+	b.BlockHeight = block.Block.Header.Height
 
 	for _, txx := range block.Block.Data.Txs {
 		var t tx.StdTx
